Added Minimum Swaps 2
- took me over an hour
- I overcomplicated the alorithm. Put to much emphasis on minumum number
of swaps. was thinking of swapping the max out of place with the min out
of place (provided the swap results in correct placement of one)
- so I looked at some of the solutions in the discussion and on the web.
they were just moving from left to right through they array. so I did
that with the C# dict lookup to get the correct index
- also read the editorial. something in there about finding cycles. I'd
like to figure that out.
- editorial also mentioned sorting the array and tracking the steps.
that is good but it would have to be the most optimal sorting algorithm in
order to get the minimum steps